Add test to prove wrong rounding in split profit

+++ packages/valory/skills/task_submission_abci/models.py
@@ -66,7 +66,8 @@ class Params(BaseParams):
         self.metadata_hash: str = self._ensure("metadata_hash", kwargs, str)
         self.task_mutable_params = MutableParams()
         self.manual_gas_limit = self._ensure_get("manual_gas_limit", kwargs, int)
-        self.service_owner_share = self._ensure("service_owner_share", kwargs, float)
+        # We are using 10_000 for 100% so 1000 == 0.1 for the service_owner_share
+        self.service_owner_share = self._ensure("service_owner_share", kwargs, int)
         self.profit_split_balance = self._ensure("profit_split_balance", kwargs, int)
         mech_to_config_dict: Dict[str, Dict[str, bool]] = self._ensure_get(
             "mech_to_config", kwargs, Dict[str, Dict[str, bool]]
